+++ lib/common/utils.h
@@ -14,6 +14,7 @@
 
 #include <cgraph/agxbuf.h>
 #include <stdbool.h>
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -43,7 +44,7 @@ typedef struct {
   unsigned utf8 : 1;
 } xml_flags_t;
 
-UTILS_API nodequeue *new_queue(int);
+UTILS_API nodequeue *new_queue(size_t);
 UTILS_API void free_queue(nodequeue *);
 UTILS_API void enqueue(nodequeue *, Agnode_t *);
 UTILS_API Agnode_t *dequeue(nodequeue *);
